feat: convert to material design
this update converts the app to material design and improves performance, use of proper material design components and concepts includes FAB and Coordinator Layout. Performance is enhanced by replacing the display of text from textview to recyclerview, which makes the app run smoothly when transitioning from one page to another.